refactor: 도메인 테스트 패키지 구조 수정
