+++ app/src/test/java/woowacourse/movie/domain/model/movie/ScreeningPeriodTest.kt
@@ -1,8 +1,7 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.movie
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import woowacourse.movie.domain.model.movie.ScreeningPeriod
 import java.time.LocalDate
 
 class ScreeningPeriodTest {
