+++ app/src/test/java/woowacourse/movie/domain/model/theater/SeatTest.kt
@@ -1,10 +1,9 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.theater
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 import woowacourse.movie.domain.model.movie.TicketType
-import woowacourse.movie.domain.model.theater.Seat
 
 class SeatTest {
     @ParameterizedTest
