+++ app/src/test/java/woowacourse/movie/domain/model/theater/SeatsTest.kt
@@ -1,10 +1,8 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.theater
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import woowacourse.movie.domain.model.movie.TicketType
-import woowacourse.movie.domain.model.theater.Seat
-import woowacourse.movie.domain.model.theater.Seats
 
 class SeatsTest {
     @Test
